backend/requirements.runtime.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
mvg==1.6.0
valkey==6.1.1
pydantic-settings==2.12.0
prometheus-client==0.23.1
SQLAlchemy==2.0.44
alembic==1.17.2
asyncpg==0.30.0
httpx==0.28.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-propagator-b3==1.38.0
